IIA_FLASK-main/requirements.txt
psycopg[binary]
google-generativeai==0.3.2

